fix: remove workaround for handling `undefined`
This workaround was added to handle an issue in `Ember.set`

https://github.com/poteto/ember-deep-set/pull/14

This original problem has long been resolved in Ember. _Not_ calling
`set` now causes some issues with Ember 3.13+, as properties on the
original object either need to be `@tracked` _or_ be updated using
`set`, the later of which makes more sense.